artur/requirements.txt
fastapi
uvicorn
Pillow
ultralytics
opencv-python
numpy
imageio
matplotlib
moviepy
argparse